Start modularizing code.
