--- data_morph/__init__.py
@@ -1,3 +1,0 @@
-"""Data Morph"""
-
-__version__ = '0.1.0'
