+++ src/data_morph/__init__.py
@@ -1,0 +1,4 @@
+"""Data Morph"""
+
+__version__ = '0.1.0'
+MAIN_DIR = __name__
