+++ src/data_morph/config/plot_style.mplstyle
